requirements.txt
fastapi>=0.111.0
uvicorn>=0.27.0
pydantic>=2.4.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
redis>=5.0.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.24.0
websockets>=12.0
numpy>=1.24.0
scipy>=1.10.0
jsonmerge>=1.9.0
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
python-dotenv>=1.0.0
httpx>=0.24.1 